Update to xlC 5.0.  (But doesn't work with optimization.)
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%402843

+++ mpqc/src/lib/chemistry/molecule/molfreq.cc
@@ -353,7 +353,7 @@ MolecularFrequencies::thermochemistry(int degeneracy, double T, double P)
 
   // compute S_el
   double S_el;
-  S_el = R*log(degeneracy);
+  S_el = R*log(double(degeneracy));
 
   // compute total molar entropy S (in J/(mol*K))
   double S;
